# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: VAKSINA HR - Kadrlarni boshqaruv dasturi
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vacancies:
    get:
      operationId: getVacancies
      tags: [vacancies]
      summary: Barcha vakansiyalar
      parameters:
        - { name: status, in: query, schema: { type: string } }
        - { name: requestId, in: query, schema: { type: number } }
        - { name: search, in: query, schema: { type: string } }
      responses:
        "200":
          description: Vakansiyalar
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vacancy"
    post:
      operationId: createVacancy
      tags: [vacancies]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VacancyInput"
      responses:
        "201":
          description: Yaratildi
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vacancy"
  /vacancies/{id}:
    get:
      operationId: getVacancy
      tags: [vacancies]
      parameters:
        - { name: id, in: path, required: true, schema: { type: number } }
      responses:
        "200":
          description: Vakansiya
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vacancy"
    patch:
      operationId: updateVacancy
      tags: [vacancies]
      parameters:
        - { name: id, in: path, required: true, schema: { type: number } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VacancyUpdate"
      responses:
        "200":
          description: Yangilandi
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vacancy"
    delete:
      operationId: deleteVacancy
      tags: [vacancies]
      summary: Ish o'rnini o'chirish (faqat HR va Direktor)
      parameters:
        - { name: id, in: path, required: true, schema: { type: number } }
      responses:
        "204":
          description: O'chirildi
        "403":
          description: Ruxsat yo'q
  /vacancies/{id}/publish:
    post:
      operationId: publishVacancy
      tags: [vacancies]
      summary: Vakansiyani kanallarga joylashtirish
      parameters:
        - { name: id, in: path, required: true, schema: { type: number } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublishVacancyInput"
      responses:
        "200":
          description: Joylashtirildi
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vacancy"

  # --- CANDIDATES (Nomzodlar) ---
...
components:
  schemas:
...
    Vacancy:
      type: object
      required: [id, requestId, title, status, createdAt]
      properties:
        id: { type: number }
        requestId: { type: number }
        title: { type: string }
        description: { type: ["string", "null"] }
        salaryRange: { type: ["string", "null"] }
        location: { type: ["string", "null"] }
        schedule: { type: ["string", "null"] }
        benefits: { type: ["string", "null"] }
        channels:
          type: array
          items:
            $ref: "#/components/schemas/VacancyChannel"
        status:
          type: string
          enum: [draft, published, closed]
        recruiterId: { type: ["integer", "null"] }
        recruiterName: { type: ["string", "null"] }
        deadline: { type: ["string", "null"] }
        requestPosition: { type: ["string", "null"] }
        departmentName: { type: ["string", "null"] }
        requestCount: { type: ["number", "null"] }
        requestRequirements: { type: ["string", "null"] }
        requestDescription: { type: ["string", "null"] }
        requestPriority: { type: ["string", "null"] }
        assignedAt: { type: ["string", "null"] }
        acceptedAt: { type: ["string", "null"] }
        publishedAt: { type: ["string", "null"] }
        createdAt: { type: string }
        candidatesCount: { type: number }
    VacancyChannel:
      type: object
      required: [channelId, channelName, channelIcon]
      properties:
        channelId: { type: number }
        channelName: { type: string }
        channelIcon: { type: string }
        publishedAt: { type: ["string", "null"] }
        views: { type: number }
        applications: { type: number }
    VacancyInput:
      type: object
      required: [requestId, title, recruiterId, deadline]
      properties:
        requestId: { type: number }
        title: { type: string }
        description: { type: string }
        salaryRange: { type: string }
        location: { type: string }
        schedule: { type: string }
        benefits: { type: string }
        recruiterId: { type: number }
        deadline: { type: string }
    VacancyUpdate:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        salaryRange: { type: string }
        location: { type: string }
        schedule: { type: string }
        benefits: { type: string }
        status: { type: string }
        recruiterId: { type: number }
        deadline: { type: string }
    PublishVacancyInput:
      type: object
      required: [channelIds]
      properties:
        channelIds:
          type: array
          items:
            type: number

    # --- CANDIDATE ---
...